build: no jmps to simpify (icw quarkus class loader)
