--- implementation/BusinessLogic/src/main/java/businesslogic/BusinessLogicAPI.java
@@ -1,12 +1,0 @@
-package businesslogic;
-
-/**
- * API of the BusinessLogic layer.
- * 
- * @author Informatics Fontys Venlo
- */
-public interface BusinessLogicAPI {
-    
-    CustomerManager getCustomerManager();
-    
-}
