--- implementation/BusinessLogic/src/main/java/businesslogic/BusinessLogicFactory.java
@@ -1,16 +1,0 @@
-package businesslogic;
-
-import persistence.PersistenceAPI;
-
-/**
- * Factory to provide BusinessLogicAPI implementation.
- * 
- * @author Informatics Fontys Venlo
- */
-public interface BusinessLogicFactory {
-    
-    static BusinessLogicAPI getImplementation( PersistenceAPI persistenceAPI ){
-        return new BusinessLogicAPIImpl( persistenceAPI );
-    }
-    
-}
