--- implementation/BusinessLogic/src/main/java/businesslogic/Customer.java
@@ -1,23 +1,0 @@
-package businesslogic;
-
-import datarecords.CustomerData;
-
-/**
- * Wrapper class that contains CustomerData and Customer Business Logic.
- * 
- * @author Informatics Fontys Venlo
- */
-public class Customer {
-    
-    private CustomerData customerData;
-
-    public Customer(CustomerData customerData) {
-        this.customerData = customerData;
-    }
-    
-    // Write Customer Business Logic (methods) here
-    // Changes to CustomerData would mean creation of a new CustomerData
-    // object and replace the existing object.
-    
-    
-}
