+++ implementation/BusinessLogic/src/main/java/io/github/fontysvenlo/ais/businesslogic/BusinessLogicAPI.java
@@ -1,0 +1,12 @@
+package io.github.fontysvenlo.ais.businesslogic;
+
+/**
+ * API of the BusinessLogic layer.
+ *
+ * @author Informatics Fontys Venlo
+ */
+public interface BusinessLogicAPI {
+
+    CustomerManager getCustomerManager();
+
+}
