+++ implementation/BusinessLogic/src/main/java/io/github/fontysvenlo/ais/businesslogic/BusinessLogicFactory.java
@@ -1,0 +1,16 @@
+package io.github.fontysvenlo.ais.businesslogic;
+
+import io.github.fontysvenlo.ais.persistence.PersistenceAPI;
+
+/**
+ * Factory to provide BusinessLogicAPI implementation.
+ *
+ * @author Informatics Fontys Venlo
+ */
+public interface BusinessLogicFactory {
+
+    static BusinessLogicAPI getImplementation( PersistenceAPI persistenceAPI ){
+        return new BusinessLogicAPIImpl( persistenceAPI );
+    }
+
+}
