--- implementation/BusinessLogic/src/main/java/module-info.java
@@ -1,6 +1,0 @@
-module businesslogic_module {   
-    requires datarecords_module;
-    requires persistence_module;
-    
-    exports businesslogic;
-}
