--- implementation/DataRecords/src/main/java/module-info.java
@@ -1,3 +1,0 @@
-module datarecords_module {       
-    exports datarecords;
-}
