+++ implementation/Persistence/src/main/java/io/github/fontysvenlo/ais/persistence/PersistenceAPIImpl.java
@@ -1,0 +1,13 @@
+package io.github.fontysvenlo.ais.persistence;
+
+/**
+ * Actual creator of storage services.
+ * @author Informatics Fontys Venlo
+ */
+class PersistenceAPIImpl implements PersistenceAPI{
+
+    @Override
+    public CustomerStorageService getCustomerStorageService() {
+        return new CustomerStorageServiceImpl();
+    }
+}
