+++ implementation/Persistence/src/main/java/io/github/fontysvenlo/ais/persistence/PersistenceFactory.java
@@ -1,0 +1,13 @@
+package io.github.fontysvenlo.ais.persistence;
+
+/**
+ * Factory to provide PersistenceAPI implementation.
+ * @author Informatics Fontys Venlo
+ */
+public interface PersistenceFactory {
+
+    static PersistenceAPI getImplementation(){
+        return new PersistenceAPIImpl();
+    }
+
+}
