--- implementation/Persistence/src/main/java/module-info.java
@@ -1,10 +1,0 @@
-module persistence_module { 
-    requires datarecords_module;
-    
-    requires java.logging;
-    requires java.sql;
-    requires org.postgresql.jdbc;
-    requires java.naming;
-    
-    exports persistence;
-}
