--- implementation/Persistence/src/main/java/persistence/CustomerStorageServiceImpl.java
@@ -1,31 +1,0 @@
-package persistence;
-
-import datarecords.CustomerData;
-import java.util.List;
-import javax.sql.DataSource;
-
-/**
- * This class knows everything about storing and retrieving customers from
- * the database. At the moment only returns dummy object with an id that is set.
- * Normally it will connect to a database and do all the handling.
- * 
- * @author Informatics Fontys Venlo
- */
-class CustomerStorageServiceImpl implements CustomerStorageService {
-
-    @Override
-    public CustomerData add(CustomerData customerData) {
-        
-        DataSource db = DBProvider.getDataSource("aisdb");
-        
-        // Create dummy customerData object with id. Normally, the id would be generated by the database
-        return new CustomerData( 1, customerData.firstName(), customerData.lastName(), customerData.dob());
-        
-    }
-
-    @Override
-    public List<CustomerData> getAll() {
-        throw new UnsupportedOperationException("Not supported yet.");
-    }
-    
-}
