--- implementation/Persistence/src/main/java/persistence/PersistenceAPIImpl.java
@@ -1,16 +1,0 @@
-package persistence;
-
-/**
- * Actual creator of storage services.
- * @author Informatics Fontys Venlo
- */
-class PersistenceAPIImpl implements PersistenceAPI{
-
-    @Override
-    public CustomerStorageService getCustomerStorageService() {
-        return new CustomerStorageServiceImpl();
-    }
-    
-    
-    
-}
