--- implementation/Persistence/src/main/java/persistence/PersistenceFactory.java
@@ -1,13 +1,0 @@
-package persistence;
-
-/**
- * Factory to provide PersistenceAPI implementation.
- * @author Informatics Fontys Venlo
- */
-public interface PersistenceFactory {
-    
-    static PersistenceAPI getImplementation(){
-        return new PersistenceAPIImpl();
-    }
-    
-}
