+++ implementation/businesslogic/src/main/java/io/github/fontysvenlo/ais/businesslogic/Customer.java
@@ -1,0 +1,23 @@
+package io.github.fontysvenlo.ais.businesslogic;
+
+import io.github.fontysvenlo.ais.datarecords.CustomerData;
+
+/**
+ * Wrapper class that contains CustomerData and Customer Business Logic.
+ *
+ * @author Informatics Fontys Venlo
+ */
+public class Customer {
+
+    private CustomerData customerData;
+
+    public Customer(CustomerData customerData) {
+        this.customerData = customerData;
+    }
+
+    // Write Customer Business Logic (methods) here
+    // Changes to CustomerData would mean creation of a new CustomerData
+    // object and replace the existing object.
+
+
+}
