+++ implementation/businesslogic/src/main/java/io/github/fontysvenlo/ais/businesslogic/CustomerManager.java
@@ -1,0 +1,32 @@
+package io.github.fontysvenlo.ais.businesslogic;
+
+import java.util.List;
+
+import io.github.fontysvenlo.ais.datarecords.CustomerData;
+import io.github.fontysvenlo.ais.persistence.CustomerStorageService;
+
+
+/**
+ * Manages customers in the business logic.
+ * Linking pin between GUI and persistence. Connected to customerStorageService
+ * in order to retrieve customers and to persist changes.
+ *
+ * @author Informatics Fontys Venlo
+ */
+public class CustomerManager {
+
+    private final CustomerStorageService customerStorageService;
+
+    public CustomerManager( CustomerStorageService customerStorageService ) {
+        this.customerStorageService = customerStorageService;
+    }
+
+    public CustomerData add( CustomerData customerData ){
+        return customerStorageService.add(customerData);
+    }
+
+    public List<CustomerData> list(){
+        return customerStorageService.getAll();
+    }
+
+}
