+++ implementation/persistence/src/main/resources/db.properties
@@ -1,0 +1,6 @@
+aisdb.dbhost=localhost
+aisdb.port=5432
+aisdb.dbname=aisdb
+aisdb.schema=ais
+aisdb.username=aisdemouser
+aisdb.password=ais
