+++ implementation/pom.xml
@@ -4,7 +4,7 @@
     <parent>
         <groupId>io.github.fontysvenlo</groupId>
         <artifactId>informaticspom</artifactId>
-        <version>1.3</version>
+        <version>1.5</version>
         <relativePath/>
     </parent>
     <groupId>io.github.fontysvenlo.ais</groupId>
@@ -13,9 +13,9 @@
     <packaging>pom</packaging>
     <name>AirlineInformationSystem</name>
     <modules>
-        <module>BusinessLogic</module>
-        <module>DataRecords</module>
-        <module>Persistence</module>
+        <module>businesslogic</module>
+        <module>datarecords</module>
+        <module>persistence</module>
         <module>restapi</module>
     </modules>
     <properties>
@@ -25,5 +25,8 @@
         <maven.compiler.source>21</maven.compiler.source>
         <maven.compiler.target>21</maven.compiler.target>
         <maven.compiler.release>21</maven.compiler.release>
+
+        <surefire-plugin.version>3.5.0</surefire-plugin.version>
+        <compiler-plugin.version>3.13.0</compiler-plugin.version>
     </properties>
 </project>
