--- implementation/restapi/src/main/java/module-info.java
@@ -1,13 +1,0 @@
-module restapi_module {
-    requires java.base;
-    requires persistence_module;
-    requires businesslogic_module;
-    requires datarecords_module;
-
-    requires jakarta.ws.rs;
-    requires quarkus.core;
-    requires jakarta.cdi;
-    requires jakarta.inject;
-
-    exports restapi;
-}
